requirements.txt

# # Create virtual environment
# python -m venv .venv

# # Activate (Windows)
# .venv\Scripts\activate

# # Install dependencies
# pip install -r requirements.txt



# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.4
pydantic-core==2.41.5
python-multipart==0.0.6
python-dotenv==1.0.0



langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
#langchain-ollama==1.0.0
#langchain-groq==1.1.0
langsmith==0.4.47



# Text Processing
nltk==3.8.1
sqlparse==0.4.4

# Vector Store / Array Support
numpy==1.26.4

# HTTP Clients
requests==2.31.0
httpx==0.28.0

# Utilities
pandas==2.1.4

# SQL Server
pyodbc==5.0.1
